refactor(params): move interp_multi_bp to opParams from debug multi-bp file
