+++ selfdrive/debug/test_multi_breakpoint.py
@@ -2,43 +2,9 @@
 
 import numpy as np
 
-from common.numpy_fast import interp, find_nearest_index, is_multi_iter
+from common.numpy_fast import interp
+from common.op_params import interp_multi_bp
 
-
-def interp_multi_bp(x, bp, v):
-  l_x = len(x)
-  l_bp = len(bp)
-  l_v = len(v)
-  is_bp_multi_iter = is_multi_iter(bp)
-  is_v_multi_iter = is_multi_iter(v)
-
-  if not is_bp_multi_iter:
-    return interp(x[-1], bp, v[-1])
-
-  if not is_multi_iter(bp[-1]):
-    bp[-1] = [bp[-1], bp[-1]]
-
-  if not is_v_multi_iter:
-    v = [v, v]
-
-  if l_v <= 1:
-    v = [v[-1], v[-1]]
-
-  if l_bp < l_x or l_bp <= 1 or len(bp[0]) <= 1:
-    # return interp(x[0], bp[0][0], v[0])
-    idx = 0
-  else:
-    idx = find_nearest_index(bp[0], x[0])
-
-  # print(f'indexes: {idx}')
-
-  if hasattr(idx, '__iter__'):
-    return [interp(x[-1], bp[-1][-1], v[i]) for i in set(idx)]
-  else:
-    return interp(x[-1], bp[-1][-1], v[idx])
-
-  # return [interp(x[-1], bp[-1][i], v[i]) for i in set(idx)] if hasattr(idx, '__iter__') else interp(x[-1], bp[-1][idx], v[idx])
-  # return interp(x[1], bp[1][idx], v[idx])
 
 def main():
   idxs = [1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1]
